DATA605/Spring2025/projects/TutorTask268_Spring2025__Analyze_Real-Time_Bitcoin_Data_with_Azure_SDK_for_Python/requirements.txt
pandas
matplotlib
python-dotenv
azure-identity
azure-eventhub
azure-storage-blob
scikit-learn
statsmodels
notebook